+++ ci/test-runner-plugin.sh
@@ -21,7 +21,6 @@ IMAGE="${IMAGE:?IMAGE variable must be set}"
 echo "==> Runner plugin tests for image: ${IMAGE}"
 echo
 
-
 # -----------------------------------------------------------------------------
 # Test 1: Plugin directory exists
 #
@@ -33,7 +32,6 @@ echo
 echo "==> Plugins: plugin directory exists"
 docker run --rm "${IMAGE}" exec test -d /usr/local/lib/runner.d
 
-
 # -----------------------------------------------------------------------------
 # Test 2: No plugins present in base image
 #
@@ -44,7 +42,6 @@ docker run --rm "${IMAGE}" exec test -d /usr/local/lib/runner.d
 
 echo "==> Plugins: no plugins present in base image"
 docker run --rm "${IMAGE}" exec sh -c '[ -z "$(ls -A /usr/local/lib/runner.d)" ]'
-
 
 # -----------------------------------------------------------------------------
 # Test completion